+++ Students/__init__.py
@@ -1,0 +1,3 @@
+from .Model import Model
+from .Dialog import Dialog
+from .View import View
